feat(A1+U1+A4): dynamic SGP denoms, configurable baselines, remove ECR self-reference
Wave 2 of Phase 1 Core Engine Unification:

A1: SGPCalculator now accepts optional `denominators` parameter. All
internal references use `self._denominators` instead of hardcoded config.
compute_sgp_denominators() can now be wired into the pipeline.

U1: LeagueConfig now has configurable roster_ab_baseline (5500),
roster_pa_baseline (6100), roster_ip_baseline (1300). Rate-stat SGP
methods use these instead of hardcoded values. Can be computed from
actual league data.

A4: Commented out heater_sgp_rank as ECR source #7. ECR consensus
now uses 6 external sources only (ESPN, Yahoo, CBS, NFBC, FanGraphs,
FantasyPros). Eliminates circular confirmation bias.

18 new tests (16 dynamic SGP + 2 ECR). 86 related tests pass.

+++ tests/test_valuation_math.py
@@ -222,9 +222,9 @@ class TestSGPRateStats:
         calc = SGPCalculator(config)
         # .300 hitter: 165 H in 550 AB
         player = _hitter(1, "Test", ab=550, h=165, avg=0.300)
-        # Code uses int() truncation: roster_h = int(5500 * 0.265) = 1457
-        roster_ab = 5500
-        roster_h = int(5500 * 0.265)  # 1457, matching code's int() truncation
+        # Code uses float math: roster_h = 5500.0 * 0.265 = 1457.5
+        roster_ab = 5500.0
+        roster_h = 5500.0 * 0.265  # 1457.5, matching code's float baseline
         new_avg = (roster_h + 165) / (roster_ab + 550)
         old_avg = roster_h / roster_ab  # 1457/5500, not 0.265
         expected = (new_avg - old_avg) / denom
@@ -238,9 +238,9 @@ class TestSGPRateStats:
         ip = 200
         er = 3.00 * ip / 9  # = 66.67
         player = _pitcher(1, "Ace", ip=ip, era=3.00, er=er)
-        # Code uses int() truncation: roster_er = int(1300 * 3.80 / 9) = 548
-        roster_ip = 1300
-        roster_er = int(1300 * 3.80 / 9)  # 548, matching code's int()
+        # Code uses float math: roster_er = 1300.0 * 3.80 / 9
+        roster_ip = 1300.0
+        roster_er = 1300.0 * 3.80 / 9  # ~548.89, matching code's float baseline
         new_era = (roster_er + er) * 9 / (roster_ip + ip)
         old_era = roster_er * 9 / roster_ip  # 548*9/1300, not 3.80
         expected = -(new_era - old_era) / denom
